Make possible to import PBassign.py
The PBassign.py script could not be imported as a module. Thus, it was
not possible to use it in an other python script and it was not possible
to write unit tests.

I added a test that execute the top level code only if the script is
called from the command line.