Makes PBlib.hclust compatible with latest R
Implement the strategies described by @hublot in issue #66. When the
ward method is selected, if R version is lesser or equal than 3.0.3,
then the squared distance matrix is passed to R hclust, if the version
is greater, then the raw matrix is passed and the method is changed to
ward.D2.

!!Warning: this commit breaks the tests as the clusters are not
identical.