Setup basic reference doc
To build the doc of the API, run `make html` in the `doc-api` directory.
The html doc is built in the doc-api/build/html directory ans index.html
is the root file.

Docstring are formated using the numpy format
<https://github.com/numpy/numpy/blob/master/doc/HOWTO_DOCUMENT.rst.txt>
or something close. This format is not handle by sphinx by default so
the docstring are nor render properly yet.

The `sphinx` python module is required to build the doc.