Add a demo notebook to the API doc
Add a notebook demonstrating how to assign PB using pbxplore API. Also
add a script that includes the notebooks in doc-api/source/notebooks to
the sphinx API doc.

The new way to build the API doc is to run `build.sh` in the doc-api
directory. That script depends on jupyter and pandoc.