Remove clustering
PB clustering, while full of potential application,
is a mostly untested feature. The lack of tests goes beyond
the absence of unit tests: indeed the scientific pertinence
and performances of the possible algorithms were not assessed.
Such assessment could be a paper by itself.

This commit removes clustering related functions and
documentation so it can come back later in all its glory.